requirements.txt
dj-database-url==0.5.0
django-debug-toolbar==2.2
django-heroku==0.3.1
Django==3.0.6
gunicorn==20.0.4
psycopg2-binary==2.8.5
whitenoise==5.2.0
asgiref==3.2.7
pytz==2019.3
sqlparse==0.3.1
